Updated code to fix more warnings. Turned on more warnings. Fixed bug where SIHttpAppBackpack was not being released due to non-obvious retains. Added TEstUtils to handle swizzling of NSURLCOnnection methods during tests.

+++ classes/PIPieman.m
@@ -102,7 +102,7 @@
 			[DDLog addLogger:[DDTTYLogger sharedInstance]];
 			server = [[HTTPServer alloc] init];
 			[server setConnectionClass:[SIHttpIncomingConnection class]];
-			[server setPort:port];
+			[server setPort:(unsigned short)port];
 			NSError *error = nil;
 			if(![server start:&error]) {
 				@throw [PIException exceptionWithReason:[NSString stringWithFormat:@"Error starting HTTP Server: %@", error]];
